requirements.txt
Django==4.2.11
lxml==5.2.1
djangorestframework==3.15.1